Refactoring - removing schemas module
Schemas module is not needed at this time. It's still not clear to me that users will stick to a regex based rule model, will they use multiple models (if so, what are they), etc. Until there is enough evidence to suggest a consistent modelling, I'll removing schemas as they reduce flexibility.